fix pagelet exports
