+++ ui/pagelets/src/dailyFeed.ts
@@ -1,4 +1,4 @@
-import flairPicker from 'common/flairPicker';
+import flairPicker from './exports/flairPicker';
 
 $('.emoji-details').each(function (this: HTMLElement) {
   flairPicker(this);
