+++ ui/pagelets/src/streamer.ts
@@ -1,6 +1,6 @@
 import * as xhr from 'common/xhr';
 import { isSafari } from 'common/device';
-import { wire as wireCropDialog } from './crop';
+import { wire as wireCropDialog } from './exports/crop';
 
 if (isSafari()) wireCropDialog(); // preload
