+++ ui/pagelets/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../tsconfig.base.json",
-  "include": ["src"],
+  "include": ["src/exports"],
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "dist",
